letta-photo-search/requirements.txt
letta>=0.13.0
uagents>=0.12.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pillow>=10.1.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=2.1.0
httpx>=0.25.0
torch>=2.9.0
transformers>=4.57.0
scikit-learn>=1.3.0
hypercorn>=0.14.0
